feat: Step별 카드로 루틴 생성 화면 구성
- Step 단위 카드 기반 루틴 생성/수정 UI 구현
- 이동 단계 및 요약 카드 포함

+++ app/src/main/java/com/smwu/bigsister/data/remote/FirestroreModels.kt
@@ -1,6 +1,8 @@
 package com.smwu.bigsister.data.remote
 
-import com.smwu.bigsister.data.local.*
+import com.smwu.bigsister.data.local.CompletionEntity
+import com.smwu.bigsister.data.local.RoutineEntity
+import com.smwu.bigsister.data.local.StepEntity
 
 data class RoutineDocument(
     val id: Long = 0L,
@@ -23,7 +25,7 @@ data class StepDocument(
     val routineId: Long = 0L,
     val orderIndex: Int = 0,
     val name: String = "",
-    val duration: Int = 0
+    val duration: Long = 0
 ) {
     constructor(entity: StepEntity) : this(
         id = entity.id,
